Add badges, Classifiers, CI/CD Testing, Improved TWST Collect, opensampl-server2 (#7)
This adds badges to the README.md for quick knowledge of the Python
software's current status, as well as classifiers to better enable
package discovery, and a CI/CD workflow for running pytest. 

Extends the Microchip TWST collect script to collect all data entries by
default and moves the filtering of which measurements should be sent to the db to 
the MicrochipTWST Probe Object. 

An additional entrypoint opensampl-server2 was added which more directly wraps 
docker compose while providing the correct compose file and env file. 
---------

Co-authored-by: Grant, Josh <grantjn@ornl.gov>
Co-authored-by: MacFarland, Midgie <macfarlandmj@ornl.gov>